Move PathGeometry from common to entity module
PathGeometry has EntityType dependencies making it entity-specific rather
than app-independent. Move it to the proper module location:

- Move path_geometry.py: common → entity
- Move test_path_geometry.py: common/tests → entity/tests
- Update import paths in EntityManager and SvgItemFactory

This maintains proper module boundaries where 'common' is reserved for
truly app-independent utilities without data model dependencies.
